requirements.txt
flask==3.0.3
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2
fpdf2==2.7.9
Pillow==10.3.0
requests==2.32.3
python-docx==1.1.2
pytesseract==0.3.13
